requirements.txt
Flask==3.0.0
prometheus-client==0.19.0
python-dotenv==1.0.0
